# api/openapi.yaml
openapi: 3.1.0
info:
  title: Simlab
  version: "1.0.0"
  description: |
    The backend of Simlab: a mining-workload simulation and emulation
    application.

    Simlab drives **runs**. A run replays a seismic-processing workload and
    asks the real autoscaler what to do about it, cycle by cycle, storing every
    observation and decision.

    Two modes, differing only in which autoscaler target the run drives:

    - **simulation** — a target on the autoscaler's `simulation` platform.
      Decisions are computed by the real decision engine and reported back;
      nothing is provisioned; time is compressed. Simlab creates the target for
      the run and removes it afterwards.
    - **live** — a target on real infrastructure. Simlab observes and records;
      it never drives, because a second controller issuing decisions would be
      two schedulers fighting over one fleet.

    This service is also the only thing that talks to the autoscaler. The
    `/api/platforms`, `/api/targets` and `/api/targets/{id}/settings` endpoints
    stand in front of it, so a browser needs no credential: the autoscaler's
    token unlocks every platform credential it holds, and a token that reaches
    a browser has been disclosed.

    Durations are seconds throughout.
servers:
  - url: /
...
paths:
...
  /api/mines:
    get:
      tags: [mines]
      summary: List mines
      responses:
        "200":
          description: Every mine, by name.
          content:
            application/json:
              schema:
                type: object
                properties:
                  mines: { type: array, items: { $ref: "#/components/schemas/Mine" } }
    post:
      tags: [mines]
      summary: Create or replace a mine
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/Mine" } } }
      responses:
        "200": { description: Saved., content: { application/json: { schema: { $ref: "#/components/schemas/Mine" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
  /api/mines/{id}:
    parameters: [{ $ref: "#/components/parameters/ID" }]
    get:
      tags: [mines]
      summary: Fetch a mine
      responses:
        "200": { description: The mine., content: { application/json: { schema: { $ref: "#/components/schemas/Mine" } } } }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [mines]
      summary: Replace a mine
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/Mine" } } }
      responses:
        "200": { description: Saved., content: { application/json: { schema: { $ref: "#/components/schemas/Mine" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
    delete:
      tags: [mines]
      summary: Delete a mine and its scenarios
      responses:
        "204": { description: Deleted. }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/scenarios:
    get:
      tags: [scenarios]
      summary: List scenarios
      parameters:
        - name: mine_id
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Scenarios.
          content:
            application/json:
              schema:
                type: object
                properties:
                  scenarios: { type: array, items: { $ref: "#/components/schemas/Scenario" } }
    post:
      tags: [scenarios]
      summary: Create or replace a scenario
      description: |
        A scenario with no `seed` is given one, because a scenario that cannot
        be replayed identically makes every comparison drawn from it
        meaningless.
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/Scenario" } } }
      responses:
        "200": { description: Saved., content: { application/json: { schema: { $ref: "#/components/schemas/Scenario" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
  /api/scenarios/{id}:
    parameters: [{ $ref: "#/components/parameters/ID" }]
    get:
      tags: [scenarios]
      summary: Fetch a scenario
      responses:
        "200": { description: The scenario., content: { application/json: { schema: { $ref: "#/components/schemas/Scenario" } } } }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [scenarios]
      summary: Replace a scenario
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/Scenario" } } }
      responses:
        "200": { description: Saved., content: { application/json: { schema: { $ref: "#/components/schemas/Scenario" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
    delete:
      tags: [scenarios]
      summary: Delete a scenario
      description: Runs that used it keep their results; their scenario reference is cleared.
      responses:
        "204": { description: Deleted. }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/runs:
    get:
      tags: [runs]
      summary: List runs, newest first
      description: |
        Each entry carries `active`, which the database alone cannot answer: a
        process that was killed leaves a run stuck at `running` in the table.
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [pending, running, completed, failed, cancelled] }
        - name: limit
          in: query
          schema: { type: integer, default: 100, maximum: 500 }
      responses:
        "200":
          description: Runs.
          content:
            application/json:
              schema:
                type: object
                properties:
                  runs:
                    type: array
                    items:
                      type: object
                      properties:
                        run: { $ref: "#/components/schemas/Run" }
                        active: { type: boolean }
    post:
      tags: [runs]
      summary: Start a run
      description: |
        Returns as soon as the run is under way; it executes in the background
        and survives the tab that started it being closed.

        A simulation run gets an autoscaler target of its own, created with the
        run and removed when it ends — including when it fails. A shared target
        would start each run holding the previous run's executors.
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/RunRequest" } } }
      responses:
        "202": { description: Started., content: { application/json: { schema: { $ref: "#/components/schemas/Run" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { description: This run is already in flight., content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /api/runs/{id}:
    parameters: [{ $ref: "#/components/parameters/ID" }]
    get:
      tags: [runs]
      summary: Fetch a run, with its metrics and the settings it ran under
      responses:
        "200":
          description: The run.
          content:
            application/json:
              schema:
                type: object
                properties:
                  run: { $ref: "#/components/schemas/Run" }
                  active: { type: boolean }
                  metrics: { $ref: "#/components/schemas/Metrics" }
                  settings: { type: object }
                  provenance: { $ref: "#/components/schemas/Provenance" }
                  reproducible:
                    type: boolean
                    description: Whether the run can be rebuilt and replayed from its provenance alone.
                  not_reproducible_because:
                    type: array
                    items: { type: string }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [runs]
      summary: Delete a run and everything recorded for it
      responses:
        "204": { description: Deleted. }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { description: The run is still in flight; cancel it first., content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /api/runs/{id}/cancel:
    parameters: [{ $ref: "#/components/parameters/ID" }]
    post:
      tags: [runs]
      summary: Stop a run
      description: Whatever it recorded before stopping stands, and it is marked cancelled rather than failed.
      responses:
        "202": { description: Stopping., content: { application/json: { schema: { type: object } } } }
        "409": { description: The run is not in flight., content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /api/runs/{id}/cycles:
    parameters:
      - $ref: "#/components/parameters/ID"
      - name: from
        in: query
        description: Return cycles after this sequence number, so a client follows a run without refetching its whole timeline.
        schema: { type: integer, default: 0 }
      - name: limit
        in: query
        schema: { type: integer, default: 5000, maximum: 20000 }
    get:
      tags: [runs]
      summary: A run's timeline
      responses:
        "200":
          description: Cycles, in order, with the sequence to resume from.
          content:
            application/json:
              schema:
                type: object
                properties:
                  cycles: { type: array, items: { $ref: "#/components/schemas/Cycle" } }
                  next: { type: integer }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /api/runs/{id}/layout:
    parameters: [{ $ref: "#/components/parameters/ID" }]
    get:
      tags: [runs]
      summary: The sensor array a run's virtual mine was replayed against
      description: |
        Recorded when the run started rather than read from the mine, which
        can be edited afterwards: this is what the autoscaler was actually
        deciding against.
      responses:
        "200":
          description: The layout.
          content:
            application/json:
              schema:
                type: object
                properties:
                  layout: { $ref: "#/components/schemas/Layout" }
        "404":
          description: |
            No such run, or a run with no virtual mine — a live run, or one
            recorded before mines were modelled. The message says which.
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  /api/runs/{id}/seismicity:
    parameters:
      - $ref: "#/components/parameters/ID"
      - name: from
        in: query
        description: Return events after this sequence number.
        schema: { type: integer, default: 0 }
      - name: limit
        in: query
        schema: { type: integer, default: 5000, maximum: 20000 }
    get:
      tags: [runs]
      summary: A run's seismic events, and when the mine could locate each one
      description: |
        Every detected event is recorded as the run starts, unlocated, and
        updated as its picks are processed. While a run is in flight, re-read
        this rather than assuming an earlier read is still current.
      responses:
        "200":
          description: Events, in order, with the sequence to resume from.
          content:
            application/json:
              schema:
                type: object
                properties:
                  events: { type: array, items: { $ref: "#/components/schemas/SeismicEvent" } }
                  next: { type: integer }
  /api/runs/{id}/entities:
    parameters: [{ $ref: "#/components/parameters/ID" }]
    get:
      tags: [runs]
      summary: The people and vehicles underground during a run, and where they went
      description: |
        Recorded as the run starts. Not paged: a day of a whole workforce's
        movement is a few megabytes, and a view needs all of it to move anyone.
      responses:
        "200":
          description: Entities, by id.
          content:
            application/json:
              schema:
                type: object
                properties:
                  entities: { type: array, items: { $ref: "#/components/schemas/Entity" } }
  /api/runs/{id}/intent:
    parameters: [{ $ref: "#/components/parameters/ID" }]
    get:
      tags: [runs]
      summary: How a run's mine reorders its queued work, and every change to that
      description: |
        `settings` and `version` are what is in force now: for a run in flight,
        the live settings, which may be a change made since the last cycle and
        not yet applied. `changes` is every version that took effect, with the
        cycle it took effect from; replaying them as `intent_schedule`
        reproduces the run.
      responses:
        "200":
          description: The run's intent.
          content:
            application/json:
              schema:
                type: object
                properties:
                  run_id: { type: string }
                  active: { type: boolean, description: Whether the run is in flight, and so whether its intent can still change. }
                  settings: { $ref: "#/components/schemas/IntentSettings" }
                  version: { type: integer }
                  schedule: { type: array, items: { $ref: "#/components/schemas/IntentStep" } }
                  changes: { type: array, items: { $ref: "#/components/schemas/IntentChange" } }
        "404":
          description: No such run, or a live run or one created before intent existed, which reordered nothing.
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    patch:
      tags: [runs]
      summary: Change an in-flight run's intent
      description: |
        A patch onto the settings in force: fields not named keep their value,
        and a list named replaces the list. It takes effect from the run's next
        cycle, never part-way through one, and is recorded with that cycle so
        the run can still be reproduced. Pass `expected_version` to make the
        write a compare-and-swap.
      parameters:
        - name: expected_version
          in: query
          schema: { type: integer }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/IntentSettings" } } }
      responses:
        "200":
          description: Accepted; in force from the next cycle.
          content:
            application/json:
              schema:
                type: object
                properties:
                  run_id: { type: string }
                  active: { type: boolean }
                  settings: { $ref: "#/components/schemas/IntentSettings" }
                  version: { type: integer }
        "400": { $ref: "#/components/responses/BadRequest" }
        "409": { description: The run is not in flight, or the intent changed since the version being edited., content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /api/runs/{id}/events:
    parameters: [{ $ref: "#/components/parameters/ID" }]
    get:
      tags: [runs]
      summary: Watch one run as it happens
      description: |
        Server-sent events. Event types are `status`, `cycle` and `metrics`.

        A watcher that falls behind loses events rather than slowing the run
        down, so a client that reconnects should re-read the run's cycles from
        `/cycles?from=` rather than assuming the stream was complete.
      responses:
        "200":
          description: An event stream.
          content:
            text/event-stream:
              schema: { type: string }
...
  /api/targets:
    get:
      tags: [autoscaler]
      summary: List the autoscaler's targets
      responses:
        "200": { description: Targets., content: { application/json: { schema: { type: object } } } }
    post:
      tags: [autoscaler]
      summary: Register a target on real infrastructure
      description: |
        Credentials go in and are never returned; the autoscaler redacts them.
        They pass through this service and are not stored here.
      requestBody:
        required: true
        content: { application/json: { schema: { type: object } } }
      responses:
        "201": { description: Registered., content: { application/json: { schema: { type: object } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
  /api/targets/{id}:
    parameters: [{ $ref: "#/components/parameters/ID" }]
    get:
      tags: [autoscaler]
      summary: Fetch a target
      responses:
        "200": { description: The target., content: { application/json: { schema: { type: object } } } }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [autoscaler]
      summary: Remove a target
      responses:
        "204": { description: Removed. }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/targets/{id}/settings:
    parameters: [{ $ref: "#/components/parameters/ID" }]
    get:
      tags: [autoscaler]
      summary: A target's live settings
      responses:
        "200": { description: Settings., content: { application/json: { schema: { type: object } } } }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [autoscaler]
      summary: Change a live autoscaler's policy
      description: |
        This is what makes the settings editor real rather than decorative: the
        change reaches the running controller and is in force on its next
        decision cycle.

        Pass `expected_version` to make the write a compare-and-swap.
      parameters:
        - name: expected_version
          in: query
          schema: { type: integer, format: int64 }
        - name: X-Simlab-Actor
          in: header
          description: Carried through to the autoscaler's change log, so the change is attributable to a person rather than to "simlab".
          schema: { type: string }
      requestBody:
        required: true
        content: { application/json: { schema: { type: object } } }
      responses:
        "200": { description: Applied., content: { application/json: { schema: { type: object } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { description: The settings changed since the version being edited., content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /api/targets/{id}/status:
    parameters: [{ $ref: "#/components/parameters/ID" }]
    get:
      tags: [autoscaler]
      summary: A target's last decision
      responses:
        "200": { description: Status., content: { application/json: { schema: { type: object } } } }
        "404": { $ref: "#/components/responses/NotFound" }

# Applies to every operation below. The probes are exempt, and say so
# individually with `security: []`.
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Required on every endpoint except the probes.

        This service is the one that is published — simlab-web's nginx proxies
        `/api` straight to it — and it holds the autoscaler's own token and
        proxies to it. An unauthenticated call here therefore reaches an
        authenticated autoscaler, and through it whatever infrastructure a
        registered target points at.

        Set `SIMLAB_API_TOKEN` to require it. Left empty the API is open, which
        only suits a development install that nothing can reach.
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema: { type: string }
  responses:
    BadRequest:
      description: The request cannot be carried out, with the reason given.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
  schemas:
...
    Mine:
      type: object
      properties:
        id: { type: string, example: storhall }
        name: { type: string }
        sensors:
          type: integer
          minimum: 1
          description: |
            How many seismic sensors the site runs. It sets the scale of
            everything downstream: more sensors means more picks per event, and
            more picks means more jobs.
        background_rate_per_hour:
          type: number
          description: Events per hour when nothing unusual is happening.
        layout:
          $ref: "#/components/schemas/Layout"
          description: |
            Where the sensors are. Optional: a mine that states none gets one
            derived from its sensor count and id, the same in every scenario.
            When stated it must place exactly `sensors` sensors, each inside
            its extent.
        description: { type: string }
        created_at: { type: string, format: date-time }
      required: [id, name, sensors]
...